Improvements on UpdateData
- When updating, a temporary folder is created while downloading and preparing new files. This is then replaced when all complete, which means that users can still query companies while updating api data.

+++ OrgOppslag/UpdateData.py
@@ -1,4 +1,5 @@
 import os
+import shutil
 
 import pandas as pd
 import numpy as np
@@ -7,41 +8,53 @@ import urllib.request
 
 base_path = os.path.realpath(os.path.dirname(__file__))
 
-def delete_old_brreg_csv():
-    print("1/4: Deleting old csv files.")
-    old_csv_files = os.listdir(os.path.join(base_path,"./data"))
-    for file in old_csv_files:
-        os.remove(os.path.join(base_path,f"./data/{file}"))
-    print("1/4: DONE.")
 
 
 def download_new_brreg():
-    print("2/4: Downloading CSV containing all companies.")
+    print("1/4: Downloading CSV containing all companies.")
     with urllib.request.urlopen("https://data.brreg.no/enhetsregisteret/api/enheter/lastned/csv") as response:
         with gzip.GzipFile(fileobj=response) as uncompressed:
             file_content = uncompressed.read()
-    with open(os.path.join(base_path, f"./data/all.csv"), 'wb') as f:
+    with open(os.path.join(base_path, f"./data/temp/all.csv"), 'wb') as f:
         f.write(file_content)
-        print("2/4: DONE.")
+        print("1/4: DONE.")
 
 
 def separate_csv_by_firstletter():
-    print("3/4: Separating companies into files by first letter.")
-    all = pd.read_csv(os.path.join(base_path,"./data/all.csv"), low_memory = False).replace(np.NaN, None)
+    print("2/4: Separating companies into files by first letter.")
+    all = pd.read_csv(os.path.join(base_path,"./data/temp/all.csv"), low_memory = False).replace(np.NaN, None)
+    all["navn"] = all["navn"].str.upper()
     grouped = all.groupby(all.navn.str[0])
     for group in grouped:
-        pd.DataFrame(group[1], columns=["organisasjonsnummer","navn"]).to_csv(os.path.join(base_path, f"./data/{ord(group[0].lower())}.csv"), index=False)
+        pd.DataFrame(group[1], columns=["organisasjonsnummer","navn"]).to_csv(os.path.join(base_path, f"./data/temp/{ord(group[0].lower())}.csv"), index=False)
+    print("2/4: DONE.")
+
+
+def delete_old_brreg_csv():
+    print("3/4: Deleting old csv files.")
+    old_csv_files = os.listdir(os.path.join(base_path,"./data"))
+    for file in old_csv_files:
+        if file == "temp":
+            continue
+        os.remove(os.path.join(base_path,f"./data/{file}"))
     print("3/4: DONE.")
 
+
+def move_temp_to_data():
+    print("4/4: Moving files from temp into data")
+    temp_folder_files = os.listdir(os.path.join(base_path,"data","temp"))
+    for file in temp_folder_files:
+        shutil.move(os.path.join(base_path,"data","temp",file), os.path.join(base_path,"data",file))
+    os.remove(os.path.join(base_path,"data","all.csv"))
+    print("4/4: DONE")
 
 
 # Main Function
 def update_brreg_files():
-    delete_old_brreg_csv() # Delete old csv files
-    download_new_brreg() # Download csv with all companies
+    download_new_brreg() # Download csv with all companies into temp
     separate_csv_by_firstletter() # Separate companies into csv by first letter
-    os.remove(os.path.join(base_path,"./data/all.csv")) # Delete file containing all companies
-    print("4/4: CSV containing all companies deleted.")
+    delete_old_brreg_csv() # Delete old csv files
+    move_temp_to_data() # Transfer all files from temp into data
     print("DONE: Updated Company Data.")
